requirements.txt
dj-database-url==0.4.2
Django==2.0.2
gunicorn==19.7.1
psycopg2==2.7.3.2
wheel==0.30.0
whitenoise==3.3.1
reportlab==3.4.0
django-import-export==1.0.0
